requirements.txt
langchain-google-genai
langchain
pandas
python-dotenv
fpdf
pyPDF2
solara